+++ src/assembler/frontend/static_analyser/Register_usage_profile.cpp
@@ -17,6 +17,7 @@
  *  along with Viua VM.  If not, see <http://www.gnu.org/licenses/>.
  */
 
+#include <algorithm>
 #include <viua/assembler/frontend/static_analyser.h>
 
 namespace viua { namespace assembler { namespace frontend {
@@ -32,6 +33,21 @@ auto Register_usage_profile::fresh(Register const r) const -> bool {
 auto Register_usage_profile::defresh() -> void {
     fresh_registers.clear();
 }
+auto Register_usage_profile::erase_arguments(Token const t) -> void {
+    auto args_regs = std::vector<Register>{};
+    std::copy_if(
+        fresh_registers.begin()
+        , fresh_registers.end()
+        , std::back_inserter(args_regs)
+        , [](Register const& r) -> bool {
+            return r.register_set == viua::internals::Register_sets::ARGUMENTS;
+        }
+    );
+    for (auto const& each : args_regs) {
+        erase(each, t);
+        fresh_registers.erase(fresh_registers.find(each));
+    }
+}
 
 auto Register_usage_profile::define(Register const r,
                                     Token const t,
@@ -39,7 +55,9 @@ auto Register_usage_profile::define(Register const r,
     using viua::internals::Register_sets;
     if ((not in_bounds(r))
         and !(r.register_set == Register_sets::GLOBAL
-              or r.register_set == Register_sets::STATIC)) {
+              or r.register_set == Register_sets::STATIC
+              or r.register_set == Register_sets::PARAMETERS
+              )) {
         /*
          * Do not thrown on global or static register set access.
          * There is currently no simple (or complicated) way to check if such
@@ -73,7 +91,9 @@ auto Register_usage_profile::define(Register const r,
     using viua::internals::Register_sets;
     if ((not in_bounds(r))
         and !(r.register_set == Register_sets::GLOBAL
-              or r.register_set == Register_sets::STATIC)) {
+              or r.register_set == Register_sets::STATIC
+              or r.register_set == Register_sets::PARAMETERS
+              )) {
         /*
          * Do not thrown on global or static register set access.
          * There is currently no simple (or complicated) way to check if such