+++ src/programinstructions.cpp
@@ -513,26 +513,6 @@ auto Program::opframe(int_op const a, int_op const b) -> Program& {
     return (*this);
 }
 
-auto Program::opparam(int_op const a, int_op const b) -> Program& {
-    addr_ptr = cg::bytecode::opparam(addr_ptr, a, b);
-    return (*this);
-}
-
-auto Program::oppamv(int_op const a, int_op const b) -> Program& {
-    addr_ptr = cg::bytecode::oppamv(addr_ptr, a, b);
-    return (*this);
-}
-
-auto Program::oparg(int_op const a, int_op const b) -> Program& {
-    addr_ptr = cg::bytecode::oparg(addr_ptr, a, b);
-    return (*this);
-}
-
-auto Program::opargc(int_op const a) -> Program& {
-    addr_ptr = cg::bytecode::opargc(addr_ptr, a);
-    return (*this);
-}
-
 auto Program::opallocate_registers(int_op const a) -> Program& {
     addr_ptr = cg::bytecode::opallocate_registers(addr_ptr, a);
     return (*this);